requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
flake8==6.1.0
idna==3.6
markdown==3.5.1
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.1.0
pyjwt==2.8.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
